feat: add TopAppBar to the diff screen
