+++ app/src/main/java/com/github/plplmax/notifications/ui/MainScreen.kt
@@ -4,26 +4,16 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.github.plplmax.notifications.MainViewModel
 import com.github.plplmax.notifications.ui.navigation.AppNavHost
-import com.github.plplmax.notifications.ui.navigation.Routes
-import com.github.plplmax.notifications.ui.notification.NotificationTopAppBar
 import com.github.plplmax.notifications.ui.snackbar.LocalSnackbarState
 
 @Composable
 fun MainScreen(viewModel: MainViewModel) {
     val navController = rememberNavController()
     Scaffold(
-        topBar = {
-            val backStackEntry by navController.currentBackStackEntryAsState()
-            if (backStackEntry?.destination?.route == Routes.Notifications.route) {
-                NotificationTopAppBar()
-            }
-        },
         snackbarHost = { SnackbarHost(hostState = LocalSnackbarState.current) }
     ) { padding ->
         AppNavHost(
