+++ app/src/main/res/values/strings.xml
@@ -33,4 +33,5 @@
     <string name="yesterday">Yesterday</string>
     <string name="no_notifications_yet">No notifications yet</string>
     <string name="failed_to_load_notifications">Failed to load notifications</string>
+    <string name="go_back">Go back</string>
 </resources>
